Rename file to match exported default
