+++ frontend/src/ContentNavigator.jsx
