+++ frontend/src/Navigation.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import ContentNavigator from "./AppNavigation";
+import ContentNavigator from "./ContentNavigator";
 import AccountNavigator from "./AccountNavigator";
 import { AuthContext } from "./services/auth/AuthProvider";
